requirements.txt
ansible==2.9.9
appdirs==1.4.4
astroid==2.4.2
asyncio==3.4.3
asyncssh==1.18.0
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
ciscoconfparse==1.5.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
dnspython==1.16.0
future==0.18.2
idna==2.9
isort==4.3.21
Jinja2==2.11.2
junos-eznc==2.2.1
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
ncclient==0.6.7
netaddr==0.7.19
netdev==0.9.3
netmiko==3.1.1
nornir==2.4.0
ntc-templates==1.4.2
nxapi-plumbing==0.5.2
paramiko==2.7.1
passlib==1.7.2
pathspec==0.8.0
pycparser==2.20
pyeapi==0.8.3
pyIOSXR==0.53
pylint==2.5.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scp==0.13.2
six==1.15.0
textfsm==1.1.0
toml==0.10.1
transitions==0.8.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wrapt==1.12.1
xmltodict==0.12.0
yamlordereddictloader==0.4.0
